refactor(core): Additional cleanup for incremental hydration
This adds comments, removes some duplicate logic, and eliminates some unnecessary complexity of the incremental hydration core logic.
