+++ packages/core/src/hydration/blocks.ts
@@ -13,16 +13,14 @@ import {Injector} from '../di';
 import {TransferState} from '../transfer_state';
 import {removeListenersFromBlocks} from '../event_delegation_utils';
 import {cleanupLContainer} from './cleanup';
-import {TNode} from '../render3/interfaces/node';
-import {HEADER_OFFSET, TView} from '../render3/interfaces/view';
-import {DeferBlock, TDeferBlockDetails} from '../defer/interfaces';
-import {getDeferBlockDataIndex, getTDeferBlockDetails, isTDeferBlockDetails} from '../defer/utils';
+import {DeferBlock} from '../defer/interfaces';
 import {whenStable, ApplicationRef} from '../application/application_ref';
 
 /**
  * Finds first hydrated parent `@defer` block for a given block id.
  * If there are any dehydrated `@defer` blocks found along the way,
  * they are also stored and returned from the function (as a list of ids).
+ * Note: This is utilizing serialized information to navigate up the tree
  */
 export function findFirstHydratedParentDeferBlock(deferBlockId: string, injector: Injector) {
   const deferBlockRegistry = injector.get(DeferBlockRegistry);
@@ -30,8 +28,11 @@ export function findFirstHydratedParentDeferBlock(deferBlockId: string, injector
   const deferBlockParents = transferState.get(NGH_DEFER_BLOCKS_KEY, {});
   const dehydratedBlocks: string[] = [];
 
-  let deferBlock = deferBlockRegistry.get(deferBlockId) ?? null;
+  let deferBlock = deferBlockRegistry.get(deferBlockId);
   let currentBlockId: string | null = deferBlockId;
+  // at each level we check if the registry has the given defer block id
+  // - if it does, we know it was already hydrated and can stop here
+  // - if it does not, we continue on
   while (!deferBlock) {
     dehydratedBlocks.unshift(currentBlockId);
     currentBlockId = deferBlockParents[currentBlockId][DEFER_PARENT_BLOCK_ID];
@@ -42,24 +43,21 @@ export function findFirstHydratedParentDeferBlock(deferBlockId: string, injector
 }
 
 /**
- * Hydrates a defer block by block name through jsaction code paths
+ * The core mechanism for incremental hydration. This recursively triggers
+ * hydration for all the blocks in the tree that need to be hydrated and keeps
+ * track of all those blocks that were hydrated along the way.
+ *
+ * @param injector
+ * @param blockName
+ * @param onTriggerFn The function that triggers the block and fetches deps
+ * @param hydratedBlocks The set of blocks currently being hydrated in the tree
+ * @returns
  */
-let _hydrateFromBlockNameImpl: typeof hydrateFromBlockNameImpl = () => {
-  return Promise.resolve({deferBlock: null, hydratedBlocks: new Set<string>()});
-};
-
-/**
- * Hydrates a defer block by block name using non jsaction code paths
- */
-let _incrementallyHydrateFromBlockNameImpl: typeof incrementallyHydrateFromBlockNameImpl = () => {
-  return Promise.resolve();
-};
-
-async function hydrateFromBlockNameImpl(
+export async function hydrateFromBlockName(
   injector: Injector,
   blockName: string,
-  onTriggerFn: (deferBlock: any) => void,
-  hydratedBlocks: Set<string>,
+  onTriggerFn: (deferBlock: DeferBlock) => void,
+  hydratedBlocks: Set<string> = new Set(),
 ): Promise<{
   deferBlock: DeferBlock | null;
   hydratedBlocks: Set<string>;
@@ -74,13 +72,19 @@ async function hydrateFromBlockNameImpl(
     injector,
   );
   if (deferBlock && blockId) {
+    // Step 2: Add the current block to the tracking sets to prevent
+    // attempting to trigger hydration on a block more than once
+    // simulataneously.
     hydratedBlocks.add(blockId);
     deferBlockRegistry.hydrating.add(blockId);
 
+    // Step 3: Run the actual trigger function to fetch dependencies
     await onTriggerFn(deferBlock);
+
+    // Step 4: Recursively trigger, fetch, and hydrate from the top of the hierarchy down
     let hydratedBlock: DeferBlock | null = deferBlock;
     for (const dehydratedBlock of dehydratedBlocks) {
-      const hydratedInfo = await hydrateFromBlockNameImpl(
+      const hydratedInfo = await hydrateFromBlockName(
         injector,
         dehydratedBlock,
         onTriggerFn,
@@ -88,7 +92,8 @@ async function hydrateFromBlockNameImpl(
       );
       hydratedBlock = hydratedInfo.deferBlock;
     }
-    // this is going to be the wrong defer block. We need to get the final one.
+    // TODO(incremental-hydration): this is likely where we want to do Step 5: some cleanup work in the
+    // DeferBlockRegistry.
     return {deferBlock: hydratedBlock, hydratedBlocks};
   } else {
     // TODO(incremental-hydration): this is likely an error, consider producing a `console.error`.
@@ -96,58 +101,19 @@ async function hydrateFromBlockNameImpl(
   }
 }
 
-/**
- * Sets the implementation for the `retrieveDeferBlockData` function.
- */
-export function enableHydrateFromBlockNameImpl() {
-  _hydrateFromBlockNameImpl = hydrateFromBlockNameImpl;
-  _incrementallyHydrateFromBlockNameImpl = incrementallyHydrateFromBlockNameImpl;
-}
-
-export async function hydrateFromBlockName(
+export async function incrementallyHydrateFromBlockName(
   injector: Injector,
   blockName: string,
-  onTriggerFn: (deferBlock: any) => void,
-): Promise<{
-  deferBlock: DeferBlock | null;
-  hydratedBlocks: Set<string>;
-}> {
-  return await _hydrateFromBlockNameImpl(injector, blockName, onTriggerFn, new Set<string>());
-}
-
-async function incrementallyHydrateFromBlockNameImpl(
-  injector: Injector,
-  blockName: string,
-  triggerFn: (deferBlock: any) => void,
+  triggerFn: (deferBlock: DeferBlock) => void,
 ): Promise<void> {
   const {deferBlock, hydratedBlocks} = await hydrateFromBlockName(injector, blockName, triggerFn);
-  removeListenersFromBlocks([...hydratedBlocks], injector);
   if (deferBlock !== null) {
+    // hydratedBlocks is a set, and needs to be converted to an array
+    // for removing listeners
+    removeListenersFromBlocks([...hydratedBlocks], injector);
     cleanupLContainer(deferBlock.lContainer);
-    const appRef = injector.get(ApplicationRef);
-    await whenStable(appRef);
+    // we need to wait for app stability here so we don't continue before
+    // the hydration process has finished, which could result in problems
+    await whenStable(injector.get(ApplicationRef));
   }
-}
-
-export function incrementallyHydrateFromBlockName(
-  injector: Injector,
-  blockName: string,
-  triggerFn: (deferBlock: any) => void,
-): Promise<void> {
-  return _incrementallyHydrateFromBlockNameImpl(injector, blockName, triggerFn);
-}
-
-/**
- * Whether a given TNode represents a defer block.
- */
-export function isDeferBlock(tView: TView, tNode: TNode): boolean {
-  let tDetails: TDeferBlockDetails | null = null;
-  const slotIndex = getDeferBlockDataIndex(tNode.index);
-  // Check if a slot index is in the reasonable range.
-  // Note: we do `-1` on the right border, since defer block details are stored
-  // in the `n+1` slot, see `getDeferBlockDataIndex` for more info.
-  if (HEADER_OFFSET < slotIndex && slotIndex < tView.bindingStartIndex) {
-    tDetails = getTDeferBlockDetails(tView, tNode);
-  }
-  return !!tDetails && isTDeferBlockDetails(tDetails);
 }
