+++ packages/core/src/hydration/event_replay.ts
@@ -44,16 +44,12 @@ import {DeferBlock, DeferBlockTrigger, HydrateTriggerDetails} from '../defer/int
 import {triggerAndWaitForCompletion} from '../defer/instructions';
 import {cleanupDehydratedViews, cleanupLContainer} from './cleanup';
 import {hoverEventNames, interactionEventNames} from '../defer/dom_triggers';
+import {DeferBlockRegistry} from '../defer/registry';
 
 /** Apps in which we've enabled event replay.
  *  This is to prevent initializing event replay more than once per app.
  */
 const appsWithEventReplay = new WeakSet<ApplicationRef>();
-
-/**
- * A set of in progress hydrating blocks
- */
-let hydratingBlocks = new Set<string>();
 
 /**
  * A list of block events that need to be replayed
@@ -94,6 +90,9 @@ export function withEventReplay(): Provider[] {
       provide: ENVIRONMENT_INITIALIZER,
       useValue: () => {
         const injector = inject(Injector);
+        // We have to check for the appRef here due to the possibility of multiple apps
+        // being present on the same page. We only want to enable event replay for the
+        // apps that actually want it.
         const appRef = injector.get(ApplicationRef);
         if (!appsWithEventReplay.has(appRef)) {
           const jsActionMap = inject(BLOCK_ELEMENT_MAP);
@@ -118,6 +117,9 @@ export function withEventReplay(): Provider[] {
               return;
             }
 
+            // We have to check for the appRef here due to the possibility of multiple apps
+            // being present on the same page. We only want to enable event replay for the
+            // apps that actually want it.
             if (!appsWithEventReplay.has(appRef)) {
               appsWithEventReplay.add(appRef);
               appRef.onDestroy(() => appsWithEventReplay.delete(appRef));
@@ -242,36 +244,27 @@ async function hydrateAndInvokeBlockListeners(
   currentTarget: Element,
 ) {
   blockEventQueue.push({event, currentTarget});
-  if (!hydratingBlocks.has(blockName)) {
-    hydratingBlocks.add(blockName);
-    await triggerBlockHydration(injector, blockName, fetchAndRenderDeferBlock);
-    hydratingBlocks.delete(blockName);
-  }
-}
-
-export async function fetchAndRenderDeferBlock(deferBlock: DeferBlock): Promise<DeferBlock> {
-  await triggerAndWaitForCompletion(deferBlock);
-  return deferBlock;
-}
-
-async function triggerBlockHydration(
-  injector: Injector,
-  blockName: string,
-  onTriggerFn: (deferBlock: any) => void,
-) {
-  // grab the list of dehydrated blocks and queue them up
-  const {dehydratedBlocks} = findFirstHydratedParentDeferBlock(blockName, injector);
-  for (let block of dehydratedBlocks) {
-    hydratingBlocks.add(block);
-  }
-  const {deferBlock, hydratedBlocks} = await hydrateFromBlockName(injector, blockName, onTriggerFn);
+  const {deferBlock, hydratedBlocks} = await hydrateFromBlockName(
+    injector,
+    blockName,
+    fetchAndRenderDeferBlock,
+  );
   if (deferBlock !== null) {
+    // TODO(incremental-hydration): extract this work into a post
+    // hydration cleanup function
+    const deferBlockRegistry = injector.get(DeferBlockRegistry);
+    deferBlockRegistry.hydrating.delete(blockName);
     hydratedBlocks.add(blockName);
     const appRef = injector.get(ApplicationRef);
     await appRef.whenStable();
     replayQueuedBlockEvents(hydratedBlocks, injector);
     cleanupLContainer(deferBlock.lContainer);
   }
+}
+
+export async function fetchAndRenderDeferBlock(deferBlock: DeferBlock): Promise<DeferBlock> {
+  await triggerAndWaitForCompletion(deferBlock);
+  return deferBlock;
 }
 
 function replayQueuedBlockEvents(hydratedBlocks: Set<string>, injector: Injector) {
